DPP-131 code style refactoring
